investbud-core/backend/pyproject.toml
[project]
name = "macrocrypto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "apscheduler>=3.11.1",
    "chartengineer>=0.1.7",
    "dfcleaner>=0.1.3",
    "diskcache>=5.6.3",
    "dkg>=8.1.0",
    "fastapi>=0.121.3",
    "fredapi>=0.5.2",
    "httpx>=0.28.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "web3>=7.14.0",
    "x402>=0.2.1",
    "yfinance>=0.2.66",
]

[project.optional-dependencies]
ws-client = [
    "websockets>=12.0",
    "eth-account>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macrocrypto"]

